+++ helpers/PrepareRelease.ps1
@@ -1,0 +1,3 @@
+﻿& $PSScriptRoot\GenerateDocs.ps1
+& $PSScriptRoot\RunPSScriptAnalyzer.ps1
+& $PSScriptRoot\RunPester.ps1